backend/requirements.txt
astroid==2.3.2
Babel==2.7.0
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.6
django-phonenumber-field==3.0.1
djangorestframework==3.10.3
djangorestframework-gis==0.14
gunicorn==19.9.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
phonenumbers==8.10.21
postgres==2.2.2
psycopg2-binary==2.8.3
PyJWT==1.7.1
pylint==2.4.3
pylint-django==2.0.11
pylint-plugin-utils==0.6
pytz==2019.3
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
twilio==6.32.0
typed-ast==1.4.0
urllib3==1.25.6
whitenoise==4.1.4
wrapt==1.11.2
